+++ .github/workflows/test.yml
@@ -27,4 +27,4 @@ jobs:
         run: bundle exec rake standard
 
       - name: Run tests
-        run: bundle exec rake test
+        run: bundle exec rake
